refactor(frontend): dedupe todayStr() and move date utils to dates.ts (task #479)
Replace inline todayStr() definitions with formatDateForInput import from
$lib/utils/dates. Move formatDueDate and timeAgo to dates.ts for reuse.

https://todo.brooksmcmillin.com/task/479
